fix: update product add logic to fix product add errors
